pyproject.toml
[project]
name = "astrbot-plugin-bilive-all"
version = "1.0.0"
description = "B站直播全功能插件 - 开播/下播提醒、直播数据统计报告"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8.0",
    "brotli>=1.0.0",
    "pillow>=9.0.0",
    "wordcloud>=1.8.0",
    "jieba>=0.42.0",
    "matplotlib>=3.5.0",
    "scipy>=1.8.0",
    "pydantic>=2.0.0",
    "aiosqlite>=0.19.0",
]
